+++ src/Field/PseudoGroupableField.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * ACF Codifier PseudoGroupableField pseudo groupable field abstract class
+ */
+
+namespace Geniem\ACF\Field;
+
+/**
+ * Abstract class GroupableField
+ */
+abstract class PseudoGroupableField extends \Geniem\ACF\Field\GroupableField {
+}