+++ src/Group.php
@@ -448,7 +448,7 @@ class Group {
 
         // Check if the target field exists in the field group.
         if ( ! isset( $this->fields[ $target ] ) ) {
-            throw new \Geniem\ACF\Exception( 'Geniem\ACF\Field: add_field_'. $action .' can\'t find given target "'. $target .'"' );
+            throw new \Geniem\ACF\Exception( 'Geniem\ACF\Field: add_field_' . $action . ' can\'t find given target "' . $target . '"' );
         }
 
         // Make a copy of the fields array to work with.
@@ -580,11 +580,11 @@ class Group {
     }
 
     /**
-     * Export current field and sub fields to ACF compatible format
+     * Export current field and sub fields to acf compatible format
      *
      * @param boolean $register Whether the field group is to be registered.
      *
-     * @return array ACF fields
+     * @return array Acf fields
      */
     public function export( $register = false ) {
         $obj = get_object_vars( $this );
@@ -599,10 +599,8 @@ class Group {
             $fields = [];
 
             foreach ( $obj['fields'] as $field ) {
-                $sub_fields = [];
-
-                if ( $field instanceof \ Geniem\ACF\Field\Tab ) {
-                    // Get the subfields from the tab
+                if ( $field instanceof \Geniem\ACF\Field\PseudoGroupableField ) {
+                    // Get the subfields
                     $sub_fields = $field->get_fields();
                 }
 
@@ -610,18 +608,49 @@ class Group {
 
                 // Add the possibly stored subfields
                 if ( ! empty( $sub_fields ) ) {
-                    foreach ( $sub_fields as $sub_field ) {
-                        $fields[] = $sub_field->export( $register );
-                    }
+                    $exported_sub_fields = $this->export_sub_fields( $sub_fields, $register );
+
+                    $fields = array_merge( $fields, $exported_sub_fields );
 
                     unset( $sub_fields );
                 }
             }
 
             // Remove keys, ACF requires the arrays to be numbered.
-            $obj['fields'] = array_values( $fields );
+            $obj['fields'] = array_filter( array_values( $fields ) );
         }
 
         return $obj;
+    }
+
+    /**
+     * Helper function to handle possible recursive pseudo group nestings.
+     *
+     * @param array   $fields Fields to export.
+     * @param boolean $register Whether the field group is to be registered.
+     * @return array
+     */
+    private function export_sub_fields( $fields, $register ) {
+        $return = [];
+
+        foreach ( $fields as $field ) {
+            if ( $field instanceof \Geniem\ACF\Field\PseudoGroupableField ) {
+                // Get the subfields
+                $sub_fields = $field->get_fields();
+            }
+
+            $return[] = $field->export( $register );
+
+            // Add the possibly stored subfields
+            if ( ! empty( $sub_fields ) ) {
+                $exported_sub_fields = $this->export_sub_fields( $sub_fields, $register );
+
+                $return = array_merge( $return, $exported_sub_fields );
+
+                unset( $sub_fields );
+            }
+        }
+
+        return $return;
     }
 }