+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- PseudoGroup field type with which you can treat multiple fields like a group but they appear in the admin as independent.
+
 ## [1.6.0] - 2018-03-19
 
 ### Added